optimize modules for pip package and lazy fetch dependencies
