+++ .gitignore
@@ -2,28 +2,36 @@
 *.pyo
 *.class
 *.log
+
 /infra/
+localstack/infra/
+
 /node_modules/
 package-lock.json
 /nosetests.xml
+
 /.venv*
+/.coverage
 .settings/
 .project
 .classpath
-/.coverage
-node_modules/
-localstack/infra/
 .DS_Store
-/build/
-/dist/
 *.egg-info/
 .eggs/
-/target/
 *.sw*
 ~*
 *~
+
+node_modules/
+/build/
+/dist/
+/target/
+
 .idea
+
 **/obj/**
 **/bin/**
+
 !bin/docker-entrypoint.sh
 
+requirements.copy.txt
