+++ localstack/services/cloudformation/cloudformation_starter.py
@@ -17,7 +17,8 @@ from localstack.utils.aws import aws_stack
 from localstack.utils.common import short_uid
 from localstack.stepfunctions import models as sfn_models
 from localstack.services.infra import (
-    get_service_protocol, start_proxy_for_service, do_run, setup_logging, canonicalize_api_names)
+    get_service_protocol, start_proxy_for_service, do_run, canonicalize_api_names)
+from localstack.utils.bootstrap import setup_logging
 from localstack.utils.cloudformation import template_deployer
 from localstack.services.awslambda.lambda_api import BUCKET_MARKER_LOCAL
