+++ localstack/utils/common.py
@@ -16,7 +16,6 @@ import logging
 import zipfile
 import binascii
 import tempfile
-import warnings
 import threading
 import traceback
 import subprocess
@@ -56,9 +55,6 @@ DOWNLOAD_CHUNK_SIZE = 1024 * 1024
 
 # set up logger
 LOG = logging.getLogger(__name__)
-# log format strings
-LOG_FORMAT = '%(asctime)s:%(levelname)s:%(name)s: %(message)s'
-LOG_DATE_FORMAT = TIMESTAMP_FORMAT
 
 # flag to indicate whether we've received and processed the stop signal
 INFRA_STOPPED = False
@@ -333,14 +329,6 @@ def md5(string):
     return m.hexdigest()
 
 
-def in_ci():
-    """ Whether or not we are running in a CI environment """
-    for key in ('CI', 'TRAVIS'):
-        if os.environ.get(key, '') not in [False, '', '0', 'false']:
-            return True
-    return False
-
-
 def in_docker():
     return config.in_docker()
 
@@ -351,43 +339,6 @@ def has_docker():
         return True
     except Exception:
         return False
-
-
-def setup_logging():
-    # determine and set log level
-    log_level = logging.DEBUG if is_debug() else logging.INFO
-    logging.basicConfig(level=log_level, format=LOG_FORMAT, datefmt=LOG_DATE_FORMAT)
-
-    # set up werkzeug logger
-
-    class WerkzeugLogFilter(logging.Filter):
-        def filter(self, record):
-            return record.name != 'werkzeug'
-
-    root_handlers = logging.getLogger().handlers
-    if len(root_handlers) > 0:
-        root_handlers[0].addFilter(WerkzeugLogFilter())
-        if is_debug():
-            format = '%(asctime)s:API: %(message)s'
-            handler = logging.StreamHandler()
-            handler.setLevel(logging.INFO)
-            handler.setFormatter(logging.Formatter(format))
-            logging.getLogger('werkzeug').addHandler(handler)
-
-    # disable some logs and warnings
-    warnings.filterwarnings('ignore')
-    logging.captureWarnings(True)
-    logging.getLogger('boto3').setLevel(logging.INFO)
-    logging.getLogger('s3transfer').setLevel(logging.INFO)
-    logging.getLogger('docker').setLevel(logging.WARNING)
-    logging.getLogger('urllib3').setLevel(logging.WARNING)
-    logging.getLogger('requests').setLevel(logging.WARNING)
-    logging.getLogger('botocore').setLevel(logging.ERROR)
-    logging.getLogger('elasticsearch').setLevel(logging.ERROR)
-
-
-def is_debug():
-    return os.environ.get('DEBUG', '').strip() not in ['', '0', 'false']
 
 
 def is_port_open(port_or_url, http_path=None, expect_success=True):
@@ -667,11 +618,6 @@ def load_file(file_path, default=None, mode=None):
     with open(file_path, mode) as f:
         result = f.read()
     return result
-
-
-def docker_container_running(container_name):
-    container_names = re.split(r'\s+', run("docker ps --format '{{.Names}}'").replace('\n', ' '))
-    return container_name in container_names
 
 
 def to_str(obj, encoding=DEFAULT_ENCODING, errors='strict'):
