+++ requirements.txt
@@ -1,16 +1,22 @@
 # Lines annotated with "#extended-lib" are excluded
 # from the dependencies when we build the pip package
 
+# Lines annotated with "#basic-lib" are included in the basic
+# version of the pip package (i.e., for "pip install localstack").
+
+# The remaining dependencies (without markers) are only included in the
+# "full" version of the pip package (i.e., "pip install localstack[full]").
+
 airspeed==0.5.10
 # Use this version until this bug is fixed: https://github.com/awslabs/amazon-kinesis-client-python/issues/99
 git+https://github.com/whummer/amazon-kinesis-client-python@maven-retries#egg=amazon_kclpy
 # amazon-kclpy==1.5.1
 awscli>=1.14.18
 boto>=2.49.0
-boto3>=1.9.71
+boto3>=1.9.71                  #basic-lib
 botocore>=1.12.13
 coverage>=4.0.3
-docopt>=0.6.2
+docopt>=0.6.2                  #basic-lib
 elasticsearch>=6.0.0,<7.0.0
 flake8>=3.6.0
 flake8-quotes>=0.11.0
@@ -20,7 +26,7 @@ flask_swagger==0.2.12
 forbiddenfruit==0.1.3
 jsonpath-rw==1.4.0
 localstack-ext>=0.8.6
-localstack-client==0.9
+localstack-client==0.9         #basic-lib
 moto-ext>=1.3.7.1
 nose>=1.3.7
 nose-timer>=0.7.5
@@ -31,7 +37,7 @@ python-coveralls>=2.9.1
 pyyaml>=3.13,<=5.1
 requests>=2.20.0
 requests-aws4auth==0.9
-six>=1.12.0
-subprocess32-ext==3.2.8.2
+six>=1.12.0                    #basic-lib
+subprocess32-ext==3.2.8.2      #basic-lib
 xmltodict>=0.11.0
 # yappi==1.0
